chore: move recovery phrase component
