+++ components/brave_wallet_ui/components/desktop/recovery-phrase/images/frosted-glass-recovery-phrase.png
